Oops, I accidentally submitted the last commit with an unfinished message.  Here is a better one:
Separating two URLs in the URL field of DESCRIPTION with
whitespace instead of a comma (and whitespace).  Whitespace was
previously used at some point, until I switched to comma for some
reason (unfortunately not documented).  Anyway, here is a reason
to ditch the comma:

The comma is a valid character in URLs [1, Section 2.2], so I guess it
should not be used as a separator of two URLs, even though R-exts [2]
instructs us to use a comma or whitespace.  In other words, blindly
splitting the URL field at every comma is not bulletproof.  I don't
know if R (CRAN?) is smarter than that, but whitespace seems like a
more robust separator.  Whitespace is listed as one of the suggested
URI (URL) delimiters in, the others being double quotes and angle
brackets.  Technically, whitespace may have to be introduced in the
middle of a URL if its length exceeds the maximum line length of the
transport medium (file format), and it is the responsibility of the
application to detect such (rare) cases. [1, Appendix C]

[1] http://tools.ietf.org/html/rfc3986
[2] http://cran.r-project.org/doc/manuals/r-release/R-exts.html


git-svn-id: svn://scm.r-forge.r-project.org/svnroot/dplr/pkg/dplR@959 edb9625f-4e0d-4859-8d74-9fd3b1da38cb