In treeMean(), building on Mikko's last comment/commit:
* Renamed first argument to 'x' (was already named so in the
  \arguments section of the .Rd file).

I changed that to "rwl" to be consistent with other functions like detrend


* Inputs are checked.

Having treeMean run naively with no ids is probably a bad idea. I think rather than have ids=NULL be the default we should have it stop with no valid ids object. And yes, I think having the output be "rwl" class is a good idea.



* A warning is produced if there are missing tree IDs, which is
  probably a rare occasion. The series with NA IDs are now
  averaged. Is this desirable or not?  The resulting average may not
  be very useful. Prior to my changes, the function would return a
  completely NA-filled "average" for series with a missing tree ID.

I think a stop here instead.


* Default 'ids' is NULL, means one core per tree (no averaging).

I think a stop here instead.


* Row names of 'ids' (if any) are matched to column names of 'x'.

Good.

* Small optimizations were made.
* Code and examples were reformatted a bit.
* Note: Should the return value inherit class "rwl"?

Yes. I think so. I looked through the help files for read.rwl etc and see that the value in the help file just indicates that a data.frame is returned. Should the help file say that the object is class "rwl" as well as "data.frame"?


git-svn-id: svn://scm.r-forge.r-project.org/svnroot/dplr/pkg/dplR@993 edb9625f-4e0d-4859-8d74-9fd3b1da38cb