Update chron.stabilized.R
I apologize for the botched code that I posted earlier. I must have cmd-z'ed a bit too much before copying my code in here. 

The if-clause to check if winLength is odd or even was wrong. WL%%2 ==1 asks for the modulo (the rest of the division) to be one, i.e. WL to be odd, it should have been !=0 or ==1, and not !=1 as in the code I uploaded. Also, the "else" part had an error, effectively using WL+1  (i+winLength as opposed to i+winLength-1) to calculate rbar. 

It should produce the correct amount of NAs to be padded now in both odd/even cases. 

I spell-checked your warnings. 

also: the `is.int` in the begininng is not a base function. How about asking `if(winLength!=round(winLength))` instead?